Moved declared actions logic in a separate hook
